test/abi: fix excluding "-pre" tags from ABI check
Commit below introduced nullglob to autotest:
(4641aa5)test/common: Provide {ext,null}glob by default in autotest suite

Without change in this commit, this resulted in not properly excluding tags
that end with "-pre". This resulted in comparing against pre-release SPDK version.
Since that version didn't yet have spdk-abi XML files for comparison,
and sanity check were missing in confirm_abi_deps() were missing.
All shared files comparison was skipped.

This commit accomplishes the following:
- Correctly excludes SPDK tags with "-pre"
- Checks if spdk-abi contains the directory for relevant SPDK version
- Checks if any shared objects were processed

Reported-by: Jacek Kalwas <jacek.kalwas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20074
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ test/make/check_so_deps.sh
@@ -84,7 +84,7 @@ function check_header_filenames() {
 }
 
 function get_release_branch() {
-	tag=$(git describe --tags --abbrev=0 --exclude=LTS --exclude=*-pre $1)
+	tag=$(git describe --tags --abbrev=0 --exclude=LTS --exclude="*-pre" $1)
 	branch="${tag:0:6}.x"
 	echo "$branch"
 }
@@ -102,6 +102,11 @@ function confirm_abi_deps() {
 		mkdir -p $source_abi_dir
 		echo "spdk-abi has not been found at $source_abi_dir, cloning"
 		git clone "https://github.com/spdk/spdk-abi.git" "$source_abi_dir"
+	fi
+
+	if [[ ! -d "$source_abi_dir/$release" ]]; then
+		echo "Release (${release%.*}) does not exist in spdk-abi repository"
+		return 1
 	fi
 
 	echo "* Running ${FUNCNAME[0]} against the latest (${release%.*}) release" >&2
@@ -241,6 +246,10 @@ EOF
 		processed_so=$((processed_so + 1))
 	done
 	rm -f $suppression_file
+	if [[ "$processed_so" -eq 0 ]]; then
+		echo "No shared objects were processed."
+		return 1
+	fi
 	echo "Processed $processed_so objects."
 	if [[ -z $user_abi_dir ]]; then
 		rm -rf "$source_abi_dir"
